# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SMS API DOCUMENTATION",
        "description": "A simple laravel school management system for secondary schools",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/students": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Get list of all students",
                "description": "Returns list of all students",
                "operationId": "getStudents",
                "responses": {
                    "200": {
                        "description": "successfully fetched students",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateStudent"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Students"
                ],
                "summary": "Create a new student",
                "description": "Create a new student",
                "operationId": "createStudent",
                "requestBody": {
                    "description": "Create student request body",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateStudent"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CreateStudent": {
                "required": [
                    "user_id",
                    "grade_id"
                ],
                "properties": {
                    "id": {
                        "description": "id of the student",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "user id",
                        "type": "integer"
                    },
                    "grade_id": {
                        "description": "user id",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Students",
            "description": "API Endpoints for Students"
        }
    ]
}